further refactor downloading of binaries
